nert: Curated Access to TERN Environmental Raster Data

Provides curated access to gridded environmental data published by Australia's Terrestrial Ecosystem Research Network (TERN; https://www.tern.org.au/), spanning national soil attribute, soil classification, soil biodiversity, soil moisture, actual evapotranspiration, vegetation canopy height and phenology products. The package reads Cloud-Optimised GeoTIFF and virtual raster files directly from the TERN data server via GDAL '/vsicurl' range requests, streaming only the requested spatial and temporal subset for efficient data transfer. The environmental raster data are returned as terra::SpatRaster objects. A convenient helper function is also provided that provides a batch-download capability given specific locations and dates, returning the extracted data values in an analysis-ready data.table. Access to the TERN datasets using this package requires a TERN API key.

See also

Author

Maintainer: Sam Rogers sam.rogers@adelaide.edu.au (ORCID)

Authors:

Sam Rogers sam.rogers@adelaide.edu.au (ORCID)

Adam H. Sparks adamhsparks@gmail.com (ORCID)

Wasin Pipattungsakul wasin.pipattungsakul@adelaide.edu.au

Russell Edson russell.edson@adelaide.edu.au (ORCID)

Max Moldovan max.moldovan@adelaide.edu.au (ORCID)

Other contributors:

Grains Research and Development Corporation (ROR) (GRDC Project CUR2210-005OPX (AAGI-CU), UOA2301-005OPX (AAGI-AU)) [funder, copyright holder]